split game-handler into multiple files for readability
